Gene expression in TCGA tumors

Plots a heatmap of genes expression in TCGA samples (peritumoral and tumor samples when a specific tumor type is specified, or tumor samples only when tumor option is set to "all")

Usage

TCGA_expression(
  tumor = "all",
  genes = NULL,
  include_CTP = FALSE,
  units = c("TPM", "log_TPM"),
  values_only = FALSE
)

Arguments

tumor

character defining the TCGA tumor type. Can be one of "SKCM", "LUAD", "LUSC", "COAD", "ESCA", "BRCA", "HNSC", or "all" (default).

genes

character naming the selected genes. The default value, NULL, takes all CT (specific) genes.

include_CTP

logical(1) If TRUE, CTP genes are included. (FALSE by default).

units

character(1) with expression values unit. Can be "TPM" (default) or "log_TPM" (log(TPM + 1)).

values_only

logical(1). If TRUE, the function will return the expression values in all samples instead of the heatmap. Default is FALSE.

Value

A heatmap of selected CT genes expression in TCGA samples. If values_only = TRUE, TPM expression data is returned instead.

Examples

if (FALSE) { # \dontrun{
TCGA_expression(
    tumor = "LUAD", genes = c("MAGEA1", "MAGEA3"),
    units = "log_TPM")
} # }