Gene expression in fetal germ cells

Plots a heatmap of genes expression in fetal germ cells, using scRNAseq data from "Single-cell roadmap of human gonadal development" (Garcia-Alonso, Nature 2022)

Usage

fetal_germcells_expression(
  genes = NULL,
  include_CTP = FALSE,
  ncells_max = 200,
  scale_lims = NULL,
  values_only = FALSE
)

Arguments

genes

character naming the selected genes. The default value, NULL, takes all CT (specific) genes.

include_CTP

logical(1) If TRUE, CTP genes are included. (FALSE by default).

ncells_max

integer(1) Sets the number of each cell type to represent on the heatmap (these cells will be randomly selected among each cell type) (set to 200 by default). If NULL, all cells are displayed.

scale_lims

vector of length 2 setting the lower and upper limits of the heatmap colorbar. By default, the lower limit is 0, and the upper limit corresponds to the third quartile of the logcounts values.

values_only

logical(1). If TRUE, the function will return the SingleCellExperiment instead of the heatmap. Default is FALSE.

Value

A heatmap of selected CT genes expression in single cells from fetal germ cells. If values_only = TRUE, a SingleCellExperiment is returned instead.

Examples

if (FALSE) { # \dontrun{
fetal_germcells_expression(include_CTP = FALSE, ncells_max = 100)
} # }