Promoter methylation of any gene in hESC

Plots a heatmap of mean promoter methylation levels of any genes in human embryonic cell lines. WGBS methylation data was downloaded from Encode. Methylation levels in tissues correspond to the mean methylation of CpGs located in range of 1000 pb upstream and 200 pb downstream from gene TSS.

Usage

hESC_mean_methylation(
  genes = NULL,
  include_CTP = FALSE,
  values_only = FALSE,
  na.omit = TRUE
)

Arguments

genes

character naming the selected genes. The default value, NULL, takes all CT (specific) genes.

include_CTP

logical(1) If TRUE, CTP genes are included. (FALSE by default).

values_only

logical(1), FALSE by default. If TRUE, the function will return the methylation values in all samples instead of the heatmap.

na.omit

logical(1) specifying if genes with missing methylation values in some tissues should be removed (TRUE by default). Note that no gene clustering will be done when methylation values are missing.

Value

Heatmap of mean promoter methylation of any gene in hESC. If values_only = TRUE, a SummarizedExperiment cobtaining methylation values is returned instead.

Examples

if (FALSE) { # \dontrun{
hESC_mean_methylation()
} # }