Methylation-Expression correlation of any genes in TCGA samples

Plots the correlation between methylation and expression values of a gene in TCGA samples.

TCGA_methylation_expression_correlation(
  tumor = "all",
  gene = NULL,
  nt_up = 1000,
  nt_down = 200,
  min_probe_number = 3,
  include_normal_tissues = FALSE,
  values_only = FALSE
)

Arguments

tumor

character defining the TCGA tumor type. Can be one of "SKCM", "LUAD", "LUSC", "COAD", "ESCA", "BRCA", "HNSC", or "all" (default).

gene

character selected gene.

nt_up

numeric(1) indicating the number of nucleotides upstream the TSS to define the promoter region (1000 by default)

nt_down

numeric(1) indicating the number of nucleotides downstream the TSS to define the promoter region (200 by default)

min_probe_number

numeric(1) indicating the minimum number of probes (with methylation values) within the selected region to calculate its mean methylation level. Default is 3.

include_normal_tissues

logical(1). If TRUE, the function will include normal peritumoral tissues in addition to tumoral samples. Default is FALSE.

values_only

logical(1). If TRUE, the function will return the methylation and expression values in TCGA samples instead of the heatmap. Default is FALSE.

Value

A scatter plot representing for each TCGA sample, gene expression and mean methylation values of probe(s) located in its promoter region (defined as 1000 nucleotides upstream TSS and 200 nucleotides downstream TSS by default). If values_only = TRUE, methylation and expression values are returned in a tibble instead.

Details

The coefficient of correlation is set to NA if no probes are found in promoter regions or if less than 1% of tumors are positive (TPM >= 1) for the gene.

Examples

if (FALSE) { # \dontrun{
TCGA_methylation_expression_correlation("LUAD", gene = "TDRD1")
} # }