Gene expression in human embryos

Plots a heatmap of genes expression in human early embryos, from "Petropoulos" scRNAseq dataset ("Single-Cell RNA-Seq Reveals Lineage and X Chromosome Dynamics in Human Preimplantation Embryos". Petropoulos et al., Cell 2016) or from "Zhu" scRNAseq dataset ("Single-cell DNA methylome sequencing of human preimplantation embryos". Zhu et al. Nat genetics 2018)

embryo_expression(
  dataset = c("Petropoulos", "Zhu"),
  genes = NULL,
  include_CTP = FALSE,
  scale_lims = NULL,
  values_only = FALSE
)

Arguments

dataset

character. Indicates which scRNAseq dataset to use. Either Petropoulos or Zhu, no default.

genes

character nameing the selected genes. The default value, NULL, takes all CT (specific) genes.

include_CTP

logical(1) If TRUE, CTP genes are included. (FALSE by default).

scale_lims

vector of length 2 setting the lower and upper limits of the heatmap colorbar. By default, the lower limit is 0, and the upper limit corresponds to the third quartile of the logcounts values.

values_only

logical(1). If TRUE, the function will return the SingleCellExperiment instead of the heatmap. Default is FALSE.

Value

A heatmap of selected CT genes expression in single cells from embryos. If values_only = TRUE, a SingleCellExperiment is returned instead.

Examples

if (FALSE) { # \dontrun{
embryo_expression(dataset = "Petropoulos", include_CTP = FALSE)
embryo_expression(dataset = "Zhu", include_CTP = FALSE)
} # }