Subset databases

Check the presence of the genes in the database then subsets the database to only keep these genes' data.

subset_database(variable = NULL, data, include_CTP = FALSE)

Arguments

variable

character() containing the names genes to keep in the data. The default value, NULL, takes CT (specific) genes

data

Summarized Experiment or SingleCellExperiment object with valid variable names.

include_CTP

logical(1) If TRUE, CTP genes are included. (FALSE by default).

Value

A Summarized Experiment or SingleCellExperiment object with only the variables data

Examples

  CTexploreR:::subset_database(variable = "MAGEA1", data = CTdata::GTEX_data())
#> see ?CTdata and browseVignettes('CTdata') for documentation
#> loading from cache
#> see ?CTdata and browseVignettes('CTdata') for documentation
#> loading from cache
#> class: SummarizedExperiment 
#> dim: 1 32 
#> metadata(0):
#> assays(1): TPM
#> rownames(1): ENSG00000198681
#> rowData names(5): external_gene_name GTEX_category q75_TPM_somatic
#>   max_TPM_somatic ratio_testis_somatic
#> colnames(32): Testis Ovary ... Uterus Vagina
#> colData names(0):